requirements.txt
langchain>=0.1.0
langchain-community
langchain-huggingface
faiss-cpu # atau faiss-gpu
sentence-transformers
transformers
torch
streamlit
python-dotenv
PyMuPDF